uint48, reuse index in memory

+++ src/adapters/interfaces/IMidnightAdapter.sol
@@ -51,10 +51,10 @@ interface IMidnightAdapter is IAdapter, IBuyCallback, ISellCallback, IRatifier {
 
     function asset() external view returns (address);
     function totalAssets() external view returns (uint128);
-    function lastUpdate() external view returns (uint40);
-    function nextMaturityFloor() external view returns (uint40);
+    function lastUpdate() external view returns (uint48);
+    function nextMaturityFloor() external view returns (uint48);
     function currentGrowth() external view returns (uint128);
-    function pendingMaturities(uint256) external view returns (uint40);
+    function pendingMaturities(uint256) external view returns (uint48);
     function availableMaturities() external view returns (uint8);
     function MAX_PENDING_MATURITIES() external view returns (uint8);
     function midnight() external view returns (address);
